feat: update deployment modules and remove MockUSDC integration
- Replaced MockUSDC with ValidatorSale1USDC and ValidatorNFTBase in deployment scripts.
- Added new ValidatorSale1USDC contract for NFT sales using USDC, including treasury management.
- Removed MockUSDC contract and related deployment module to streamline the codebase.
- Updated package.json scripts to reflect new deployment configurations for ValidatorNFT and ValidatorSale1USDC.
