+++ contracts/package.json
@@ -21,9 +21,8 @@
         "hh:deploy:qrtest": "npx hardhat ignition deploy ./ignition/modules/QRCodeTest.ts --network sepolia --verify",
         "hh:deploy:validatornft:base": "npx hardhat ignition deploy ./ignition/modules/ValidatorNFT.ts --network base --verify",
         "hh:deploy:validatorsale:base": "npx hardhat ignition deploy ./ignition/modules/ValidatorSale.ts --network base --verify",
-        "hh:deploy:mockusdc:basesepolia": "npx hardhat ignition deploy ./ignition/modules/MockUSDC.ts --network baseSepolia --verify",
-        "hh:deploy:validatornft:basesepolia": "npx hardhat ignition deploy ./ignition/modules/ValidatorNFT.ts --network baseSepolia --verify",
-        "hh:deploy:validatorsale:basesepolia": "npx hardhat ignition deploy ./ignition/modules/ValidatorSaleTestnet.ts --network baseSepolia --verify"
+        "hh:deploy:validatornft1usdc:base": "npx hardhat ignition deploy ./ignition/modules/ValidatorNFTBase.ts --network base --verify",
+        "hh:deploy:validatorsale1usdc:base": "npx hardhat ignition deploy ./ignition/modules/ValidatorSale1USDC.ts --network base --verify"
     },
     "devDependencies": {
         "@nomicfoundation/hardhat-chai-matchers": "^2.0.0",
